rootston: cleanup for multiseat
