+++ rootston/xcursor.c
@@ -1,14 +1,6 @@
 #include <wlr/types/wlr_cursor.h>
 #include "rootston/input.h"
 
-struct wlr_xcursor *get_default_xcursor(struct wlr_xcursor_theme *theme) {
-	return wlr_xcursor_theme_get_cursor(theme, "left_ptr");
-}
-
-struct wlr_xcursor *get_move_xcursor(struct wlr_xcursor_theme *theme) {
-	return wlr_xcursor_theme_get_cursor(theme, "grabbing");
-}
-
 static const char *get_resize_xcursor_name(uint32_t edges) {
 	if (edges & ROOTS_CURSOR_RESIZE_EDGE_TOP) {
 		if (edges & ROOTS_CURSOR_RESIZE_EDGE_RIGHT) {
@@ -30,6 +22,14 @@ static const char *get_resize_xcursor_name(uint32_t edges) {
 		return "w-resize";
 	}
 	return "se-resize"; // fallback
+}
+
+struct wlr_xcursor *get_default_xcursor(struct wlr_xcursor_theme *theme) {
+	return wlr_xcursor_theme_get_cursor(theme, "left_ptr");
+}
+
+struct wlr_xcursor *get_move_xcursor(struct wlr_xcursor_theme *theme) {
+	return wlr_xcursor_theme_get_cursor(theme, "grabbing");
 }
 
 struct wlr_xcursor *get_resize_xcursor(struct wlr_xcursor_theme *theme,
